thus

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,5 +1,4 @@
 #include <stdio.h>
-#include <stdlib.h>
 #include "function_pointers.h"
 /**
  * array_iterator - function that executes a function
@@ -10,7 +9,7 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	size_t i;
+	unsigned int i;
 
 	if (array != NULL && size > 0 && action != NULL)
 	{
